performance-evaluation/requirements.txt
addict==2.4.0
aio-pika==6.8.1
aiodns==3.0.0
aiohttp==3.8.1
aioredis==2.0.1
aiormq==3.3.1
aioschedule==0.5.2
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.25.0
atlastk==0.13
attrs==21.4.0
autopep8==1.6.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.0.0
cartooner==0.0.2
cartoonizer==0.0.4
-e git+https://github.com/DiverseSolutions/ccxt.git@a4aa3db4dd0d6d2ddbaab99534ef0daff4f36957#egg=ccxt&subdirectory=python
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
ConfigArgParse==1.5.3
cryptography==36.0.1
Cython==0.29.32
fastapi==0.75.2
fastapi-utils==0.2.1
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
frozenlist==1.2.0
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
h11==0.13.0
httptools==0.4.0
idna==3.3
install==1.3.5
iso8601==0.1.16
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-import==0.2.2
locust==2.15.1
MarkupSafe==2.1.2
mmcv==1.6.2
msgpack==1.0.5
multidict==5.2.0
mysql-connector-python==8.0.29
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pamqp==2.3.0
Pillow==9.2.0
protobuf==4.22.3
psutil==5.9.5
psycopg2==2.9.3
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
pypika-tortoise==0.1.3
python-dotenv==0.19.2
python-slugify==6.0.1
pytimeparse==1.1.8
pytz==2021.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.27.1
retrying==1.3.3
roundrobin==0.0.4
scipy==1.9.1
simplejson==3.19.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
sqlite3-to-mysql==1.4.16
starlette==0.17.1
tabulate==0.9.0
text-unidecode==1.3
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tortoise-orm==0.17.8
tqdm==4.65.0
typing_extensions==4.0.1
Unidecode==1.3.6
urllib3==1.26.8
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
Werkzeug==2.3.3
yapf==0.32.0
yarl==1.7.2
zope.event==4.6
zope.interface==6.0
